+++ backend/src/resource/asset/index.html
@@ -9,6 +9,6 @@
             .tag-btn{color:white;border-radius:4px;text-shadow:0 1px 1px rgba(0, 0, 0, 0.2);background:rgb(28, 184, 65);margin-right:6px}
             .tag-link{margin-right:6px}
         </style>
-    <script type="module">import init from '/asset/index-a6f73cf3c48f50dc.js';init('/asset/index-a6f73cf3c48f50dc_bg.wasm');</script></head>
+    <script type="module">import init from '/asset/index-665b7454336b637.js';init('/asset/index-665b7454336b637_bg.wasm');</script></head>
     <body>
 </body></html>
