+++ backend/src/resource/asset/snippets/blog-frontend-3607fd663da6040e/asset/editor.js
@@ -1,0 +1,13 @@
+export let editor = null;
+export function initEditor() {
+    const Editor = toastui.Editor;
+    editor = new Editor({
+        el: document.querySelector('#editor'),
+        previewStyle: 'vertical',
+        initialEditType: 'wysiwyg',
+        height: '500px',
+    });
+}
+export function getContent() {
+    return editor.getMarkdown();
+}
