+++ backend/src/service/asset_list.rs
@@ -1,6 +1,7 @@
 [
-("index-a6f73cf3c48f50dc.js", include_bytes!(r#"..\resource\asset/index-a6f73cf3c48f50dc.js.gz"#), "text/javascript"),
-("index-a6f73cf3c48f50dc_bg.wasm", include_bytes!(r#"..\resource\asset/index-a6f73cf3c48f50dc_bg.wasm.gz"#), "application/wasm"),
+("index-665b7454336b637.js", include_bytes!(r#"..\resource\asset/index-665b7454336b637.js.gz"#), "text/javascript"),
+("index-665b7454336b637_bg.wasm", include_bytes!(r#"..\resource\asset/index-665b7454336b637_bg.wasm.gz"#), "application/wasm"),
 ("index.html", include_bytes!(r#"..\resource\asset/index.html.gz"#), "text/html; charset=utf-8"),
 ("purecss.min-b565cdb90b7aac9f.css", include_bytes!(r#"..\resource\asset/purecss.min-b565cdb90b7aac9f.css.gz"#), "text/css"),
+("snippets/blog-frontend-3607fd663da6040e/asset/editor.js", include_bytes!(r#"..\resource\asset/snippets\blog-frontend-3607fd663da6040e\asset\editor.js.gz"#), "text/javascript"),
 ]
